tests/wpt/web-platform-tests/css/css-align/animation/justify-no-interpolation.html

<!DOCTYPE html>
<link rel=author href="mailto:jarhar@chromium.org">
<link rel=help href="https://drafts.csswg.org/css-align/#justify-items-property">
<link rel=help href="https://drafts.csswg.org/css-align/#justify-self-property">
<link rel=help href="https://github.com/w3c/csswg-drafts/issues/4441">
<link rel=help href="https://drafts.csswg.org/css-transitions-2/#transition-property-property">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/interpolation-testcommon.js"></script>

<body>
<script>
test_no_interpolation({
  property: 'justify-items',
  from: 'initial',
  to: 'baseline'
});

test_no_interpolation({
  property: 'justify-self',
  from: 'initial',
  to: 'baseline'
});
</script>
